requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
rembg==2.0.50
Pillow==10.2.0
onnxruntime==1.16.3
